refactor breaking change detector
